@@ -1,10 +1,10 @@
 source 'https://rubygems.org'
 
-ruby '2.4.9'
+ruby '2.5.8'
 
-gem 'rails', '~> 5.2.4'
+gem 'rails', '~> 6.0.0'
 
-gem 'activerecord-postgis-adapter', '~> 5.0'
+gem 'activerecord-postgis-adapter', '~> 6.0'
 gem 'aws-sdk-elasticbeanstalk'
 gem 'aws-sdk-rails'
 gem 'bootsnap', require: false
@@ -37,7 +37,7 @@ gem 'kaminari'
 gem 'newrelic_rpm'
 gem 'pg'
 gem 'puma'
-gem 'rails_admin', '~>1.4.2'
+gem 'rails_admin', '~>2.0.0'
 gem 'responders'
 gem 'sass-rails', '~>5.0'
 gem 'simple_form'
